Rename api module to reflect it is no longer abstract
